Move Node to separate file, add GetNode method
