+++ GenericsHomework/GenericsHomework/CircularSinglyLinkedList.cs
@@ -2,56 +2,7 @@
 
 public class CircularSinglyLinkedList<T>
 {
-    public class Node
-    {
-        public T Data
-        {
-            get => _Data!;
-            set => _Data = value ?? throw new ArgumentNullException(nameof(Data));
-        }
-        private T? _Data;
-
-        public Node Next
-        {
-            get => _Next!;
-            private set => _Next = value ?? throw new ArgumentNullException(nameof(Next));
-        }
-        private Node? _Next;
-        public Node(T data)
-        {
-            Data = data;
-            Next = this;
-        }
-        public void Append(Node head, T data)
-        {
-            Node temp = head;
-            while (temp.Next != head)
-            {
-                temp = temp.Next;
-            }
-            Node newNode = new(data);
-            temp.Next = newNode;
-            newNode.Next = head;
-        }
-        public override string ToString()
-        {
-            return Data!.ToString()!;
-        }
-
-        public bool Exists(Node head, T data)
-        {
-            Node temp = head;
-            while (temp.Next != head)
-            {
-                if (temp.Data!.Equals(data)) return true;
-                else
-                    temp = temp.Next;
-            }
-            return temp.Data!.Equals(data);
-        }
-    }
-
-    public Node? Head { get; set; }
+    public Node<T>? Head { get; set; }
 
     public int Size { get; private set; }
 
@@ -69,15 +20,14 @@ public class CircularSinglyLinkedList<T>
 
     public void Append(T data)
     {
-        //TODO: Throw Error if Appending duplicate value.
         if (Head is null) 
         {
-            Head = new Node(data);
+            Head = new Node<T>(data);
             Size++;
         }
         else if (Exists(data))
         {
-            throw new ArgumentException(nameof(data));
+            throw new ArgumentException("Given data already exists in the list", nameof(data));
         }
         else
         {
@@ -88,16 +38,21 @@ public class CircularSinglyLinkedList<T>
 
     public T Get(int index)
     {
-        if(Head is null)
+        return GetNode(index).Data;
+    }
+
+    public Node<T> GetNode(int index)
+    {
+        if (Head is null)
             throw new ArgumentOutOfRangeException(nameof(index));
-        Node currentNode = Head;
-        for(int i = 0; i < index; i++)
+        Node<T> currentNode = Head;
+        for (int i = 0; i < index; i++)
         {
             currentNode = currentNode.Next;
-            if(currentNode == Head)
+            if (currentNode == Head)
                 throw new ArgumentOutOfRangeException(nameof(index));
         }
-        return currentNode.Data;
+        return currentNode;
     }
 
     public void Clear(T data)
